+++ android/capacitor/src/main/java/com/getcapacitor/CapacitorWebView.java
@@ -1,12 +1,17 @@
 package com.getcapacitor;
 
 import android.content.Context;
+import android.os.Build;
 import android.util.AttributeSet;
+import android.util.TypedValue;
 import android.view.KeyEvent;
 import android.view.inputmethod.BaseInputConnection;
 import android.view.inputmethod.EditorInfo;
 import android.view.inputmethod.InputConnection;
 import android.webkit.WebView;
+import androidx.core.graphics.Insets;
+import androidx.core.view.ViewCompat;
+import androidx.core.view.WindowInsetsCompat;
 
 public class CapacitorWebView extends WebView {
 
@@ -48,5 +53,37 @@ public class CapacitorWebView extends WebView {
             return false;
         }
         return super.dispatchKeyEvent(event);
+    }
+
+    public void edgeToEdgeHandler(Bridge bridge) {
+        String configEdgeToEdge = bridge.getConfig().adjustMarginsForEdgeToEdge();
+
+        if (configEdgeToEdge.equals("disable")) return;
+
+        boolean autoMargins = false;
+        boolean forceMargins = configEdgeToEdge.equals("force");
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.VANILLA_ICE_CREAM && configEdgeToEdge.equals("auto")) {
+            TypedValue value = new TypedValue();
+            boolean foundOptOut = getContext().getTheme().resolveAttribute(android.R.attr.windowOptOutEdgeToEdgeEnforcement, value, true);
+            boolean optOutValue = value.data != 0; // value is set to -1 on true as of Android 15, so we have to do this.
+
+            autoMargins = !(foundOptOut && optOutValue);
+        }
+
+        if (forceMargins || autoMargins) {
+            ViewCompat.setOnApplyWindowInsetsListener(this, (v, windowInsets) -> {
+                Insets insets = windowInsets.getInsets(WindowInsetsCompat.Type.systemBars() | WindowInsetsCompat.Type.displayCutout());
+                MarginLayoutParams mlp = (MarginLayoutParams) v.getLayoutParams();
+                mlp.leftMargin = insets.left;
+                mlp.bottomMargin = insets.bottom;
+                mlp.rightMargin = insets.right;
+                mlp.topMargin = insets.top;
+                v.setLayoutParams(mlp);
+
+                // Don't pass window insets to children
+                return WindowInsetsCompat.CONSUMED;
+            });
+        }
     }
 }