+++ ios/Capacitor/Capacitor/CAPBridgeProtocol.swift
@@ -59,6 +59,8 @@ import WebKit
     func eval(js: String)
     // swiftlint:enable identifier_name
 
+    @objc optional func injectScriptBeforeLoad(path: String)
+
     func triggerJSEvent(eventName: String, target: String)
     func triggerJSEvent(eventName: String, target: String, data: String)