+++ ios/Capacitor/Capacitor/WebViewDelegationHandler.swift
@@ -47,7 +47,6 @@ open class WebViewDelegationHandler: NSObject, WKNavigationDelegate, WKUIDelegat
         bridge?.reset()
     }
 
-    @available(iOS 15, *)
     open func webView(
         _ webView: WKWebView,
         requestMediaCapturePermissionFor origin: WKSecurityOrigin,
@@ -58,7 +57,6 @@ open class WebViewDelegationHandler: NSObject, WKNavigationDelegate, WKUIDelegat
         decisionHandler(.grant)
     }
 
-    @available(iOS 15, *)
     open func webView(_ webView: WKWebView,
                       requestDeviceOrientationAndMotionPermissionFor origin: WKSecurityOrigin,
                       initiatedByFrame frame: WKFrameInfo,
@@ -158,6 +156,8 @@ open class WebViewDelegationHandler: NSObject, WKNavigationDelegate, WKUIDelegat
     }
 
     open func webViewWebContentProcessDidTerminate(_ webView: WKWebView) {
+        CAPLog.print("⚡️  WebView process terminated")
+        bridge?.reset()
         webView.reload()
     }